requirements.txt
numpy~=1.24.2
matplotlib~=3.7.1
torch
pyaudio~=0.2.13
scipy~=1.7.3
gym~=0.26.2
torchvision~=0.15.0
torchviz~=0.0.2
pillow~=9.4.0
sympy~=1.11.1
keras~=2.11.0